Implement remapping of cached build scripts
Build scripts now use 2 distinct caches: one uses a file hash + compile classpath key and generates "raw" build script classes, that are shared between builds (even if not in the same project). The second cache takes a script path, locates a corresponding class file in the first cache and remaps the class file so that it looks as if it was compiled locally, introducing source file location.